Move immutable module so tests can pick it up
