+++ app/javascript/mastodon/utils/immutable.ts
